Avoid using preprocessor directives
